Ensure all classes are detected in the first place
Add a test that disables the detection of class modifier exception—and thus reports all classes even `enum`/`annotation`/`sealed`/… ones—in order to validate that our RegEx detects all classes in the first place (i.e. regardless of the various possible variations of the Kotlin syntax that can be used around declaring a class) and we won't miss any due to a syntax we didn't anticipate.